+++ src/constants/account-constant.js
@@ -21,14 +21,22 @@ const COMMERCIO_COLOR = Object.freeze({
   },
 });
 
+const TRANSACTION_TYPES = Object.freeze({
+  CW20_RECEIVED: 'cw20-received',
+  CW20_SENT: 'cw20-sent',
+  RECEIVED: 'received',
+  SENT: 'sent',
+});
+
+const EVENT_FILTERS = Object.freeze({
+  [TRANSACTION_TYPES.CW20_RECEIVED]: 'wasm.to',
+  [TRANSACTION_TYPES.CW20_SENT]: 'wasm.from',
+  [TRANSACTION_TYPES.RECEIVED]: 'transfer.recipient',
+  [TRANSACTION_TYPES.SENT]: 'transfer.sender',
+});
+
 export default Object.freeze({
-  TRANSACTIONS_NUMBER: 10,
-  TRANSACTION_TYPES: Object.freeze({
-    NATIVE_SENT: 'native-sent',
-    NATIVE_RECEIVED: 'native-received',
-    CW20_SENT: 'cw20-sent',
-    CW20_RECEIVED: 'cw20-received',
-  }),
+  EVENT_FILTERS,
   MEMBERSHIPS: [
     {
       name: COMMERCIO_COLOR.GREEN.NAME,
@@ -51,4 +59,6 @@ export default Object.freeze({
       color: COMMERCIO_COLOR.BLACK.HEX,
     },
   ],
+  TRANSACTIONS_NUMBER: 10,
+  TRANSACTION_TYPES,
 });